pyproject.toml
[project]
name = "soundcloud-organizer"
version = "0.1.0"
description = "A CLI tool to automatically organize your SoundCloud stream into monthly playlists."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "typer",
    "loguru",
    "rich",
    "requests",
    "requests-oauthlib",
    "pydantic",
    "python-dateutil",
]

[project.scripts]
soundcloud-organizer = "soundcloud_organizer.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = ["pytest", "pytest-mock"]
